requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
openpyxl==3.1.2
pillow==10.0.0
pytesseract==0.3.10
aiofiles==23.1.0
requests==2.31.0
google-generativeai
pillow